Decision Integration V5 N Final ?

+++ Xi0nGUI/src/decisional/StateMachineTransitionForDecisionV2.java
@@ -36,7 +36,7 @@ public class StateMachineTransitionForDecisionV2 {
     // ------------------------------------
 	
 	private int rightSideDistance;
-	private int frontalDistance;
+	private float frontalDistance;
 	private float servoAngle;
 	
 	// ------------------------------------
