+++ Xi0nGUI/src/decisional/StateMachineTransitionForDecisionV3.java
@@ -37,7 +37,7 @@ public class StateMachineTransitionForDecisionV3 {
     // ------------------------------------
 	
 	private int rightSideDistance;
-	private int frontalDistance;
+	private float frontalDistance;
 	private float servoAngle;
 	
 	// ------------------------------------
