+++ Xi0nGUI/src/decisional/StateMachineTransitionForDecisionV4.java
@@ -39,7 +39,7 @@ public class StateMachineTransitionForDecisionV4 {
     // ------------------------------------
 	
 	private int rightSideDistance;
-	private int frontalDistance;
+	private float frontalDistance;
 	private float servoAngle;
 	
 	// ------------------------------------
